requirements.txt
httpx~=0.16.1
discord.py
